fix sealos verion github

+++ .github/scripts/apps.sh
@@ -1,10 +1,4 @@
 #!/bin/bash
-#version=${1:-v3.22.1}
-#registry=${2:-docker.io}
-#repo=${3:-cuisongliu}
-#username=${4:-cuisongliu}
-#password=${5:-}
-#app=${6:-calico}
 prefix=$registry/$repo
 buildDir=.build-image
 mkdir -p $buildDir
@@ -18,6 +12,10 @@ fi
 sh init.sh amd64
 [ -f init.sh  ] && cp  sh init.sh amd64
 sudo sealos build -t $prefix/$app:$version-amd64 --platform linux/amd64 -f $filename  .
+if [ $? != 0 ]; then
+   echo "====build app image failed!===="
+   exit 1
+fi
 cd ../ && rm -rf $buildDir
 mkdir -p $buildDir
 cp -rf applications/$app/$version/* $buildDir/
@@ -25,7 +23,10 @@ cp -rf applications/$app/$version/* $buildDir/
 cd $buildDir
 [ -f init.sh  ] && cp  sh init.sh arm64
 sudo sealos build -t $prefix/$app:$version-arm64 --platform linux/arm64 -f $filename  .
-
+if [ $? != 0 ]; then
+   echo "====build app image failed!===="
+   exit 1
+fi
 sudo buildah login --username $username --password $password $registry
 sudo buildah push $prefix/$app:$version-amd64
 sudo buildah push $prefix/$app:$version-arm64
@@ -33,4 +34,7 @@ sudo buildah manifest create $prefix/$app:$version
 sudo buildah manifest add $prefix/$app:$version docker://$prefix/$app:$version-amd64
 sudo buildah manifest add $prefix/$app:$version docker://$prefix/$app:$version-arm64
 sudo buildah manifest push --all $prefix/$app:$version docker://$prefix/$app:$version
-echo "script $prefix/$app:$version build successfully!"
+if [ $? != 0 ]; then
+   echo "====push app image failed!===="
+   exit 1
+fi
